launcher: do not consider a skipped activation job as successful
systemd jobs are currently skipped if the unit is not startable, in the sense
that it doesn't have a concept of 'starting', such as a device unit.
While it makes sense for generic interfaces like systemctl to avoid considering
this a hard failure, in the case of service activation via D-Bus this is less
sensible as only service units can be used like this, and they are all startable.

In the future we might also make service units job return 'skipped' when a
Condition* fails, instead of 'done'. If that happens, the launcher will be
able to return a sensible error to the caller, instead of just hanging, waiting
for a unit that will never be started because its activation job was skipped.

Fixes #276

See also: https://github.com/systemd/systemd/issues/21520

Signed-off-by: Luca Boccassi <luca.boccassi@microsoft.com>