launcher: use new StartUnitWithFlags method if available
Check if the new method from systemd is available via Introspect(),
and if so use it.
This allows us to receive 'skipped' as the result when the service we
are trying to activate does not run because of a failed Condition
check, which would return a false 'done' with the older
StartUnit method, resulting in us just waiting for a unit to
show up that will never arrive.

Signed-off-by: Luca Boccassi <luca.boccassi@microsoft.com>