launch/service: Check for ConditionResult on service unit
When a unit doesn't satisfy its conditions, systemd will report that the
unit was started but don't actually start the service. This cause bus
activation to hang when a service can't be started due to unsatisfied
conditions.

Fix this by explicitly checking for the ConditionResult property on the
service unit object.