launch/service: check for ConditionResult on service unit
When a unit doesn't satisfy its conditions, systemd will report that the
unit was started but don't actually start the service. This causes bus
activation to hang when a service can't be started due to unsatisfied
conditions.

Fix this by explicitly checking for the ConditionResult property on the
service unit object.

Original-by: msizanoen1 <msizanoen@qtmlabs.xyz>
(adjusted some minor coding-style issues and added comments)
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>