bus/peer: fix missing peer-object charges
We happened to use `user` rather than `peer->user` as user object to
charge new peers on. This will not work as expected as `user` is cleared
to NULL just few lines above. Meaning, peer objects are not charged at
all (but their operations are!).

Fix this and properly charge on `peer->user`.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>