util/string: avoid unsigned overflow
For some reason, gcc with -O3 warns about the `while (n--)` loop
complaining that the loop writes beyond the buffer boundaries. With a
normal for-loop this does not happen. Switch over for the sake of
keeping gcc happy.

I could not see any UB in the previous implementation, so this sounds
like a false-positive to me. But I might be wrong. The for-loop is
definitely easier to read.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>