util/user: keep reference to user in each usage table
Keep a reference to an owning user in each usage table. We want to allow
callers to hold charges without holding on to any user references.

Also fix the peer-deinitialization to be ordered correctly and free the
user references last (in particular, after the charges). This is not
strictly necessary, but now follows our coding style and would have
avoided possible failures.

This fixes an assertion failure when disconnecting entire groups of
peers of the same user, due to the recent fix that actually made
peer-accounting do something.

Reported-by: Torge Matthies
Reported-by: Mel34
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>