launch/config: avoid parsing bus type
The bus type is usually "session" or "system", but can also be
a different arbitrary string. Avoid parsing it into an enum, so
that any non-standard bus type is allowed.

dbus-broker does not use the bus type itself, but it is propagated
to AppArmor security checks.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>
(make payload-names match enum-names; fix bus-type destructor)
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>