test/osi: drop unused and broken mkosi tests
The `mkosi` tool requires package-management tools to be installed on
the host. This is often not the case and cumbersome to use. Drop it for
now, since we have a replacement with the fedora container.

Note that the current mkosi arch-based configuration was long broken,
since it cannot properly setup the ArchLinux GPG keys due too overlong
paths. This means, no-one has been using this for years, anyway.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>