bus/apparmor: always update bustype on policy update
Make sure to always propagate the new bustype on policy updates to the
apparmor code to reflect the actual policy and let the AppArmor code
decide on what to do, rather than hard-coding the behavior outside.

Note that right now policy_registry_import() is only ever called on a
fresh policy object (and does not really work otherwise), but lets
pretend this can be used to update policies as well, just to hard-code
less of this behavior.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>