bus/driver: introduce DRIVER_E_UNIMPLEMENTED for stubs
Introduce the new DRIVER_E_UNIMPLEMENTED error-code for stub functions.
This avoids returning EPERM to users and makes it easier to recognize.

Note that previously EPERM was used to keep compatibility to dbus-daemon
and avoid clients reacting to other error-codes. EPERM was a suitable
and reasonable way to emulate a possibly misconfigured bus, rather than
a custom error code. However, for the debug-calls this is definitely not
necessary and only a remnant of when less driver calls were supported.

Suggested-by: draconicfae
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>