launch/service: retain active unit name
Store the name of the unit we spawned for activation in the service
object and retain it for as long as the activation is valid.
Additionally, remember whether this is a transient unit.

This fixes a bug where a unit-reload while an activation is pending
would possibly cause a change of the unit-name that is tracked. While
it is unclear whether a change of service-files during runtime is
supported, this at least fixes any possible runtime issues and correctly
delays the unit-name update for when the activation is retried.

This also prepares for planned reworks of the spawn-handling by
remembering the name of the unit that is tracked and thus allows delayed
activation after a match-rule has been established.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>