launch/service: keep service data in separate object
We retain all information from a service file in the Service object.
This data, however, might get updated when the service file is re-read.

Separate the service-data from the runtime management data of the
Service object by moving it into its own type. This makes it better
visible which data is updated at which point in time.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>