bus/peer: fix catching OOM when linking match-rules
When the match-rule infrastructure was extended to have per-key lookup
tables, all call-sites had to be amended to catch errors when linking
new rules. One such site was missed, so fix this and catch the error
properly.

The conversion was initially performed in:

    commit 6c2b082522a6af033cd6699b669a4de614c41ad7
    Author: Tom Gundersen <teg@jklm.no>
    Date:   Wed Apr 25 17:16:47 2018 +0200

        match: make match_rule_link() fallible

This fixes a possible bug where ENOMEM is ignored and thus a match rule
is silently dropped instead of being linked and retained.

Reported-by: Mark Esler <mark.esler@canonical.com>
Signed-off-by: David Rheinsberg <david@readahead.eu>