launcher: start unit immediately as unprivileged user
It is better to never have privileges rather than start with them and
remove them later, as the attack surface is reduced, and there are
fewer things to do before being 'ready'. Nowadays systemd can run the
service as the appropriate user/group out of the box.

When starting as root files in /proc/self/fdinfo/ will be owned as root
and set to 400, so we cannot read them. Nowadays it is not necessary to
start as root when running under systemd, so just add User/Group with
the configured user to the system unit. Use the canonical 'messagebus'
user/group, if it needs to be changed a drop-in can be used.

If libaudit support is enabled, add AmbientCapabilities=CAP_AUDIT_WRITE
so that we can still write to the audit log. Make it optional depending
on the build config to ensure the additional privilege is granted only
if it can be actually used.

Signed-off-by: Luca Boccassi <bluca@debian.org>