launch: list directories alphabetically
When enumerating directories ensure that we iterate the entries in
alphabetical order.

We used to follow what the reference-implementation does, and simply
read directories in semi-random on-disk order. This is rather brittle
to debug and exposes behavior that we really don't want to expose.
Instead, we want predictable enumerations so policy and service-files
can be more easily traced and debugged.

The downside of ordering directory entries is that we have to collect
all entries first, rather than using the streaming API of the kernel.
Fortunately, this is what we do, anyway, since we already have 1-to-1
mappings of the directory entries to internal data-structures. Hence,
in all our current use-cases, it is not an issue to stream everything
into a collection first.

Signed-off-by: David Rheinsberg <david@readahead.eu>