build: order executable targets
Add `link_depends:` arguments to Meson targets which require other
executables to be built. Otherwise, `meson test` will not build the
required executables.

While `test()` targets can list other targets as dependency, this is not
possible for `executable()` or `declare_dependency()` targets. Instead,
we use the `link_depends:` argument to get a similar effect. Ideally,
this would not be required, but there seems to be not other way to
declare dependencies on other executables right now.

Signed-off-by: David Rheinsberg <david@readahead.eu>