launch: drop error-checking remnant of old dir-iteration
When the directory-iteration was changed to use fs_dir_list(), the loop
was changed to no longer check for `errno`. Unfortunately, the trailing
`errno`-check was not dropped. This means any `errno`-leftovers from the
loop will cause the entire operation to fail.

This can certainly happen when we handle syscall/libc failures
gracefully. In particular, this breaks when stale or broken service
files are left on the system.

Fix this and drop the error-checking remnant.

Reported-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>
Reported-by: Lily Danzig <ldenardo@dealerinspire.com>
Signed-off-by: David Rheinsberg <david@readahead.eu>