test: integration tests
Let's follow-up on 0a0b0fa and introduce a second Packit job that runs
integration tests on the just built RPMs in Testing Farm [0].

The test organization and execution is handled by tmt [1]. To not
unnecessarily pollute the dbus-broker's git root, the tmt root is
initialized in the test/integration/ subdirectory (meaning that for the
tmt commands to work correctly you need to be somewhere below this
directory).

Since there are no pre-existing integration tests, add a simple one that
runs dfuzzer on the org.freedesktop.systemd1 D-Bus interface to
exercise dbus-broker a bit. A short README file is provided as well with
instructions on how to easily add a new test case, and with some links
where to find more information about tmt.

[0] https://packit.dev/docs/configuration/upstream/tests
[1] https://tmt.readthedocs.io/en/stable/overview.html