build: install_tests -> tests
Change the configuration option to include tests in the distribution to
just `tests`. We try to never conditionalize which parts to build, but
rather always build everything that can be built with the selected
dependencies. Hence, a `tests` configuration would never imply that it
builds tests, and we can safely use it to mean the tests will be part of
the distribution.

There are some exceptions to this rule, but those we likely cannot
change for compatibility reasons.

Signed-off-by: David Rheinsberg <david@readahead.eu>