ci: rebase on archlinux
Switch the base image again from Fedora to Arch Linux. This requires us
to disable SELINUX in the unit-tests, but those will be run by the
packit integration, so we should be fine.

The advantages are:

 * Image builds take *significantly* less time: 10min for Fedora, 3min
   for Arch Linux (on the same infrastructure).

 * Images are significantly smaller in size: 6GiB for Fedora, 3GiB for
   ArchLinux (with the same packages).

We also get rolling releases, allowing us to control at which point we
can update. We do not expose the CI system to untrusted inputs, anyway
(nor do we pass sensitive data to it).

Signed-off-by: David Rheinsberg <david@readahead.eu>