ci: major rework
This is a major rework of the continuous integration. The aim is to
centralize any code build & analysis hooks in a single workflow, and
call other workflows from it. This allow serializing them, if desired.

In particular, this re-introduces the `ci.yml` workflow, which is meant
to be the main workflow run on pushes or PRs. It first runs a set of
basic tests, which should catch most simple mistakes found in PRs, like
typos or simple build failures. Only if those succeed, it will spawn the
extended test-suite.

This setup conserves CI resources at minimal cost. The basic tests run
really fast, and thus the experience for most PRs stays the same. But we
avoid running the test matrix for simple mistakes like missing commit
hunks, typos, failed force-pushes, etc. Given that the CI resources are
limited for the `bus1` organization, this should overall increase CI
performance, because machines become available earlier.

As a second step, this PR extends the CI considerably with the following
new features:

- By default, we now use CLANG with LLD for better performance and
  code analysis.
- We make use of native meson support for `debug`, `werror`, and
  `errorlogs`, ensuring those values are properly passed along.
- We run `clippy` on the Rust code-base for static analysis.
- We use `ubuntu` based containers for most runs to ensure we can build
  with both SELinux *and* AppArmor. No other distro allows this.
- We add runs which use native ARM64 machines to test on non-x86
  hardware. This requires container images for arm64, which was
  implemented separately.
- We invoke the CodeQL workflow from within this workflow, to ensure it
  is also guarded by the basic checks.
- We now run some checks with LTO and thinLTO.
- Instead of the `-m32` hack, we now use proper cross-compilation via
  meson cross files for native 32-bit tests.
- We now run UBSAN and ASAN in CI.
- Valgrind integration is reworked and now run as proper individual
  test.

Signed-off-by: David Rheinsberg <david@readahead.eu>