ci: use explicit i686 run
Avoid putting the i686 run into the matrix and instead use a separate
run so it is fully under our control. It differs from the standard runs,
so lets not try to force it into the same matrix setup.

For now, also add a custom cross-file, given that the one shipped with
upstream distros does not support bindgen and rustdoc.

Signed-off-by: David Rheinsberg <david@readahead.eu>