build: pull in libc-rs as dependency
Pull in `libc-rs` as dependency, as it will be needed by the session
binary. We provide our own meson integration via Meson `packagefiles`,
given that Meson Cargo wraps are unstable (neither do they support
`build.rs`).

The libc crate is straightforward to support, given that it has no
dependencies (and very likely never will), and its `build.rs` is just an
evaluation of which `cfg` options to set. And there, the options
relevant for linux can be statically defined.

Signed-off-by: David Rheinsberg <david@readahead.eu>