units: drop Sockets= declaration
Drop the explicit Sockets= declaration in unit files, to ensure we do
not share listener sockets. This was added in:

commit 954362424aad6f31283c8960901e88221066b4cd
Author: Tom Gundersen <teg@jklm.no>
Date:   Mon Aug 28 21:05:32 2017 +0200

    units: order dbus-broker.service before basic.target

However, the intention was to make use of the automatic dependency
declarations when `Sockets=` is used. Since then, we have switched to
explicit dependency configuration, so this is no longer needed.

Moreover, with it, you can run `dbus-broker.service` *and*
`dbus.service` simultaneously, and they will share the listener socket,
leading to all kinds of errors. Lets not support that, and ensure that
the system and user instances only work if used via their alias.

Signed-off-by: David Rheinsberg <david@readahead.eu>