launcher: import new metrics socket
Provide an optional additional socket for the system and user bus,
called dbus-metrics.socket. Unlike dbus.socket, it is optional and it is
shipped by us (rather than by a meta dbus package). If the unit is
enabled, it will be picked up by dbus-broker and used to export metrics.

For now, the launcher searches the inherited file-descriptors for it and
passes it to the internal launcher representation. It is not made any
further use of for now.

Signed-off-by: David Rheinsberg <david@readahead.eu>