[TEST] make sure CIFuzz works
it's just a test to make sure that CIFuzz works with
Ubuntu 24.04 with https://github.com/google/oss-fuzz/pull/14245.

It was made to work in https://github.com/google/oss-fuzz/pull/14382 and
I haven't seen any issues with other projects yet but I'll trigger it
once here just in case.